Implemented and tested bottomUpNodes(). Fixed a traversal bug
there was a redundant bottom-up, top-down, bottom-up at m_endTreeAddr. That interfered with the summation.
Specifically, node values were added correctly on the first bottom-up sweep, up to level 1; but in the second bottom-up sweep,
children were not marked dirty, so all the copies of nodes for last-children were overwritten with zeros before
accumulating all the way to level 0.