Add the readNodeNonhangingIn() property to tree loop. Plan for intergridTransfer.
My thought is that during interpolations we should treat cg nodes once per node.
The simplest way to do this is handle it outside the traversal, just set a node
to zero if it is the duplicate case or if it doesn't exist in the fine grid.
I'm not sure this is correct, so I've left it commented out for now.