Option to accumulate in MatvecBaseOut::bottomUpNodes, or not. Set to not.
I have to set to Not in both intergridTransfer and setDiag, and then multigrid seems to converge after one vcycle.

If I only set to Not for intergridTransfer but keep accumulation in setDiag, then multigrid diverges.
That doesn't seem right.

Also I have yet to try 3D and 4D and higher order.