Apply MPI_Comm_free fixes from maksbh
I applied the changes by hand since the changes were recorded as commits in another repo.

Also I used a different solution to correctly free Comms in distCoalesceSiblings().
This function repeatedly splits and discards Comms in a while loop until
the partition meets a criterion (namely, siblings are on the same mpi rank).
In my implementation of the fix, I only free each temporary comm if is != the input comm.