Trash GlobalSubset due to orphaned ghost nodes problem...
Orphaned Ghost Nodes: When an element is selected that ghost-reads a remote node,
but on the process owning the node, no incident element is selected.
The node is not selected on the owning process. The ghost node is orphaned
because the canonical instance of the node is not owned by any rank.