Modify dollar to allow function names with semicolons (like templates)
Old method identifies a frame by a concatenation of titles, delimited by semicolons,
and identifies the lineage of a frame by tokenizing the deimited string.
New method does not concatenate or tokenize, but maps each title to a unique id,
and represents the stack trace as a std::vector of numeric ids.