fixup serialize: Compensate for compiler bugs RE eval order
Expressions in a braced-init-lilst are supposed to be evaluated from left to right
(unlike arguments to a function, which the compiler is free to evaluate in any order).
However, sometimes the compilers don't obey this rule.
To maintain the ordering, the braced-init-list can be replaced by a recursive function template.