Add synch points for nbx to avoid race condition, rewrite sendAll().
The synch points are necessary to avoid interference on MPI_Recv(MPI_ANY_SOURCE).
While this could be avoided with careful management of tags,
we have no control over what tags the client code would use after this function.
So, settle for latency-hiding nonblocking synchronization,
with maximum flexibility in when to synch peers and when to synch globally.