Document why coarsen test warns: 2:1-balance too conservative
It was meant for when we use the global sort version, we might split octants across a void chasm.
But this version of the code also balances by adding parent-level neighbors of siblings
(neighbors of the parent).
So that is why the concrete example fails to coarsen one island of octants.