Do not assume pointer stability of octlist in feMat
Instead, get the octlist on-demand from the da, which we can assume is pointer-stable.
In turn, da accesses the distTree, which it assumes (and checks in debug mode) to be pointer-stable.
The current implementation of distTree holds multiple octlists as a vector of vectors;
if the outer vector grows later (by adding to the hierarchy), pointers to inner vectors may be invalidated.