(wip) Add mg Jacobi smoother, restriction, setup in driver, and use mg as preconditioner in cg
next is to add prolongation and coarse grid solver.

right now the restriction is calculated but not used.
therefore the effect of the vcycle is 2 iterations of damped jacobi,
so we can see the effect of jacobi preconditioning of cg
(it reduced the number of iterations by a small percentage).