Modify nonhanging vertex generator to handle quadratic nodes also.
Note: Vertex ownership is not enough to infer facet ownership. Consider this example,
where vertices prefer the cells below, then the cells to the left.
The lower facet of the middle element is incorrectly identified as borrowed,
because all vertices on the facet are borrowed:

     _ _
    |_|_|_
       ^|_|

A previous attempted implementation, that erroneously relies on vertex-based
inferrence is also included in the commented-out lines.