Clean up data members of Point class -> only std::array
This removes reliance on member references pointing to array entries.

Also, because of the in-class member initializer for the array,
constructors and assignment operators can be defaulted.

Finally, the equality comparison operator is fixed so that points
now compare by double-precision coordinates. The old behavior
first cast entries to int, probably assuming large integer coordinates;
however, the way the Point class is currently used, the nearest integers
are always 0 or 1, so rounding is not appropriate here.