Begin storing elemental matrices in Eigen for hybrid MatVec.
Works for a uniform grid, but not yet for nonuniform grids.

Probably due to redundant interpolations in setup and matvec().
During setup, collectMatrixEntries() builds in the interpolation
and dependence of hanging nodes from parent nodes.
During matvec(), the hanging nodes are passed in, instead of
the parent nodes that the stored matrix is expecting.